+++ src/main/java/teammates/ui/webapi/request/FeedbackResponseCommentCreateRequest.java
@@ -1,0 +1,11 @@
+package teammates.ui.webapi.request;
+
+/**
+ * The create request of a feedback response comment.
+ */
+public class FeedbackResponseCommentCreateRequest extends FeedbackResponseCommentBasicRequest {
+
+    public FeedbackResponseCommentCreateRequest(String commentText, String showCommentTo, String showGiverNameTo) {
+        super(commentText, showCommentTo, showGiverNameTo);
+    }
+}
