+++ src/main/java/teammates/ui/webapi/request/FeedbackResponseCommentUpdateRequest.java
@@ -1,39 +1,11 @@
 package teammates.ui.webapi.request;
 
-import javax.annotation.Nullable;
-
 /**
  * The update request of a feedback response comment.
  */
-public class FeedbackResponseCommentUpdateRequest extends BasicRequest {
-
-    private String commentText;
-    @Nullable
-    private String showCommentTo;
-    @Nullable
-    private String showGiverNameTo;
+public class FeedbackResponseCommentUpdateRequest extends FeedbackResponseCommentBasicRequest {
 
     public FeedbackResponseCommentUpdateRequest(String commentText, String showCommentTo, String showGiverNameTo) {
-        this.commentText = commentText;
-        this.showCommentTo = showCommentTo;
-        this.showGiverNameTo = showGiverNameTo;
+        super(commentText, showCommentTo, showGiverNameTo);
     }
-
-    public String getCommentText() {
-        return commentText;
-    }
-
-    public String getShowCommentTo() {
-        return showCommentTo;
-    }
-
-    public String getShowGiverNameTo() {
-        return showGiverNameTo;
-    }
-
-    @Override
-    public void validate() {
-        // TODO decide what to validate
-    }
-
 }
